# Improve performance of InputTGLF()