--- a/Project.toml
+++ b/Project.toml
@@ -19,7 +19,7 @@ StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 [compat]
 BSON = "0.3"
 Flux = "0.14, 0.15, 0.16"
-GACODE = "1"
+GACODE = "1.0.4"
 IMAS = "2, 3, 4, 5"
 IMASutils = "1.5.1"
 Measurements = "2"