This improves allocations. 

### Performance
For an individual run, it's like a 10% reduction in allocations and time
`@btime FUSE.TGLFNN.InputTGLF($eqt, $cp1d, $gridpoint_cp, $sat, $electromagnetic, $lump_ions)`
- Old: `97.041 μs (3017 allocations: 164.49 KiB)`
- New: `90.208 μs (2695 allocations: 151.41 KiB)`

But this gets called inside the FluxMatcher optimizer, so the whole actor it's more like 17%:
```
function g()
    @checkout :init dd ini act
    FUSE.ActorFluxMatcher(dd, act);
end
@btime g()
```
- Old: `791.279 ms (10744925 allocations: 754.97 MiB)`
- New: `663.126 ms (8944953 allocations: 629.40 MiB)`

Overall for a run of
```
function f()
           dd = IMAS.dd()
           ini, act = FUSE.case_parameters(:D3D, :default)
           FUSE.init(dd, ini, act)
           act.ActorStationaryPlasma.max_iterations = 5
           FUSE.ActorStationaryPlasma(dd, act)
           return dd
       end
@btime f()
```
- Old: `4.869802 seconds (33.22 M allocations: 2.562 GiB, 6.24% gc time)`
- New: `4.150545 seconds (24.25 M allocations: 1.903 GiB, 5.14% gc time)`

### Accuracy
For an individual run, it gives very close to the same inputTGLF results. The differences I see are:
```
BETAE
  Union{Missing, Float64}[0.014603695179053237, 0.013646685298689431, 0.010624576088015888, 0.007065409234463239, 0.004206405051703228, 0.002366544831349765, 0.0012006194630162185]
  Union{Missing, Float64}[0.014603695179053237, 0.013646685298689431, 0.010624576088015888, 0.007065409234463239, 0.0042064050517032275, 0.0023665448313497646, 0.0012006194630162185]
P_PRIME_LOC
  Union{Missing, Float64}[-0.001628990452469532, -0.001319207698012358, -0.004032558721623881, -0.0030727475428486075, -0.0021846331899535052, -0.001628321662569855, -0.0012725100133884728]
  Union{Missing, Float64}[-0.001628990452469532, -0.0013192076980123848, -0.004032558721623889, -0.0030727475428485996, -0.0021846331899535035, -0.0016283216625698523, -0.00127251001338847]
Q_PRIME_LOC
  Union{Missing, Float64}[3.6673672243717768, 4.1900096229392, 5.4202391370526, 6.008275026931598, 11.266735442885338, 19.95307360831579, 43.82738882311746]
  Union{Missing, Float64}[3.667367224371777, 4.1900096229392, 5.4202391370526, 6.008275026931597, 11.266735442885338, 19.95307360831579, 43.82738882311747]
```
So `BETAE` and `Q_PRIME_LOC` vary at machine precision. `P_PRIME_LOC` is at the 10^-14 or 10^-15 level. I think it's a combination of the change to `bunit` and moving a constant factor outside of the gradient for some reason.