@bclyons12 2025-05-19
Two important things to note:

1. Requires https://github.com/ProjectTorreyPines/IMAS.jl/pull/260
2. I made a bunch of small commits as I was debugging the changes in the result, since the convergence percentage after three iterations of ActorStationaryPlasma was changing even do the the very smaller changes in InputTGLF I documented above. We should be sure to squash this when it gets merged.