@bclyons12 2025-05-21
@tomneiser Similar changes should be done for `InputCGYRO` as well, but I'm not going to hold up this PR for that.